refactor(chat): separate the turn loop from the world it runs in
`chat/` had become a flat pile: the control flow, its policies, the agent
registry, the permission rulesets, the prompt text and the Durable Object all
sat side by side, and reading "how does a turn progress?" meant knowing which of
nine files to ignore.

The loop and its policies move to `chat/loop/`, one question per file:

  turn.ts      the control flow — the only file that decides anything
  types.ts     the vocabulary: what goes in, what comes out, one step's state
  budgets.ts   every ceiling, together
  retry.ts     which failures earn another attempt
  context.ts   keeping the request inside the model's window
  delegate.ts  handing a sub-question to a sub-agent, which re-enters turn.ts
  index.ts     the barrel callers import, so the split stays free to move

What stays outside is the world the loop runs in, not part of how it progresses:
`ChatSession.ts` (log, ordering, turn slot), `turn-runner.ts` (bootstrap),
`agents.ts` and `permissions.ts` (config), `prompts.ts`, and a new `tools.ts`
holding `buildChatTools` and `submit_diagnosis` — the loop decides *when* to call
a tool, not which tools exist.

The one non-mechanical change is `budgets.ts`. Those constants were spread across
three modules, where each looked independently reasonable and their product did
not: MAX_STEPS (10) x TASK_BUDGET_PER_TURN (4) x SUBAGENT_MAX_STEPS (6) is 240
model calls, which has to stay under TURN_STALE_MS or the DO declares a running
turn abandoned. Together, that relationship is visible and TURN_STEP_BUDGET reads
as the backstop it is.

`tagged`/`turnEnd` now take the task ref structurally rather than the whole turn
input, which is what breaks the types <-> events import cycle the split would
otherwise have created.

Also fixes a dead comparison in delegate.test.ts that only `typecheck:test`
caught: `ChatTurnEvent` excludes `user-message`, so that branch was unreachable
and the assertion was weaker than it read.
