+++ apps/api/src/chat/loop/context.ts
