+++ apps/api/src/chat/loop/index.ts
@@ -1,0 +1,25 @@
+/**
+ * The chat turn loop.
+ *
+ * Split out of the surrounding `chat/` directory so the control flow can be read on its own. The
+ * rest of `chat/` is the world the loop runs in — the Durable Object that owns the log and the turn
+ * slot, the agent registry, the permission rulesets, the prompt text, the tool set. None of that
+ * decides how a turn *progresses*; all of it is input.
+ *
+ * Read in this order:
+ *
+ *   1. `turn.ts`     — the control flow. Start here; it is the only file that decides anything.
+ *   2. `types.ts`    — the vocabulary: what goes in, what comes out, what one step carries.
+ *   3. `budgets.ts`  — every ceiling, together, because they multiply against each other.
+ *   4. `retry.ts`    — which failures earn another attempt.
+ *   5. `context.ts`  — keeping the request inside the model's window.
+ *   6. `delegate.ts` — handing a sub-question to a sub-agent, which re-enters `turn.ts`.
+ *
+ * Callers should import from here rather than reaching into a file, so the internal split stays
+ * free to move.
+ */
+export { runChatTurn } from "./turn"
+export { makeTurnUsage, type ChatTurnEvent, type ChatTurnInput, type TurnUsage } from "./types"
+export { MAX_STEPS, SUBAGENT_MAX_STEPS } from "./budgets"
+export { isNearContextLimit, pruneToolResults } from "./context"
+export { isRetryableStepFailure, stepRetryDelayMs } from "./retry"
