+++ apps/api/src/chat/loop/types.ts
@@ -1,0 +1,140 @@
+/**
+ * The vocabulary a turn speaks: what goes in, what comes out, and what one step carries.
+ *
+ * Separate from `turn.ts` so the loop's control flow reads as control flow. Everything here is data
+ * plus two small stamping helpers; nothing here decides anything.
+ */
+import type { ChatEvent, ChatTaskRef } from "@maple/domain/chat-session"
+import type { Message, Model, Tools, Usage } from "@maple/llm"
+import type { TenantContext } from "@/services/auth/tenant-context"
+import type { AgentDefinition } from "../agents"
+import type { StepRetryBudget, TaskBudget } from "./budgets"
+
+/** Distributive `Omit`, so each union member keeps its own shape. */
+type WithoutSeq<T> = T extends unknown ? Omit<T, "seq"> : never
+
+/**
+ * Events this turn wants appended to the session log. `seq` is assigned by the Durable Object,
+ * which owns the ordering, so the loop emits everything without one. `user-message` is excluded:
+ * the session records the user turn at submission time, before the loop runs.
+ */
+export type ChatTurnEvent = WithoutSeq<Exclude<ChatEvent, { type: "user-message" }>>
+
+export interface ChatTurnInput {
+	readonly sessionId: string
+	readonly tenant: TenantContext
+	readonly model: Model
+	/** The full transcript so far, oldest first, already including the new user message. */
+	readonly messages: ReadonlyArray<Message>
+	readonly messageId: string
+	/**
+	 * Investigate-mode sessions get a `submit_diagnosis` tool. It is supplied rather than built
+	 * inside the loop because it needs `InvestigationService`, which would otherwise drag the
+	 * service graph into the loop's imports.
+	 */
+	readonly extraTools?: Tools
+	/**
+	 * Whether this turn still holds the session's turn slot.
+	 *
+	 * Checked between steps so an abort takes effect at the next boundary instead of only after the
+	 * in-flight model call drains, and so a turn that has been superseded stops writing into a
+	 * conversation that has moved on. Defaults to "always current" for callers with no session.
+	 */
+	readonly isCurrent?: () => boolean
+	/** Accumulates this turn's token usage; see {@link TurnUsage}. */
+	readonly usage?: TurnUsage
+	/**
+	 * Which agent this turn runs as. Defaults to the primary agent the session id names, so existing
+	 * callers keep the behaviour their mode already had.
+	 */
+	readonly agent?: AgentDefinition
+	/**
+	 * Set when this turn is a sub-agent nested inside a parent turn: every event it produces is
+	 * stamped with this ref, which is what routes them into the parent's task card rather than the
+	 * top-level conversation.
+	 */
+	readonly task?: ChatTaskRef
+	/** Nesting depth. 0 is the conversation's own turn. */
+	readonly depth?: number
+	/** Shared across the parent and every descendant; see {@link TaskBudget}. */
+	readonly taskBudget?: TaskBudget
+	/**
+	 * Sink for events produced by a nested sub-agent turn.
+	 *
+	 * A side channel rather than a merge into the returned stream. `Stream.merge` would race the
+	 * parent's terminal `turn-end` against undrained child events, and `ChatSession.pump` closes the
+	 * SSE connection on a terminal event — so a child's tail could be written to the durable log
+	 * *after* the turn had been declared over. The consumer's `Stream.runForEach` is strictly
+	 * sequential, so an event emitted from inside a tool's `execute` is deterministically ordered
+	 * after the tool-call announcement that introduced it and before the tool-result that closes it.
+	 */
+	readonly emit?: (event: ChatTurnEvent) => void
+}
+
+/** Where one step sits in the turn. `attempt` is 0-based and resets per step. */
+export interface StepState {
+	readonly step: number
+	readonly attempt: number
+	readonly budget: StepRetryBudget
+	/** Resolved once at the top of the turn, so every step agrees on the ruleset and step cap. */
+	readonly agent: AgentDefinition
+	/** Shared with every descendant sub-agent turn. */
+	readonly taskBudget: TaskBudget
+	/**
+	 * This is the tool-less step that closes a turn which ran out of steps. Its natural exit is
+	 * "no tool calls", which would otherwise report `"stop"` and lose the signal the client badges
+	 * on — so the reason is carried here instead.
+	 */
+	readonly closing?: true
+}
+
+/**
+ * Running token total for one turn, accumulated across its steps.
+ *
+ * Mutable and shared rather than returned, because the one consumer — `submit_diagnosis` — is a
+ * *tool* invoked mid-turn, so there is no "after the turn" moment at which to hand it a total.
+ * In practice the diagnosis call is the last thing an investigation does, so this is the whole turn
+ * bar the final assistant message. Before this, `SubmitDiagnosisRequest` was built with no usage at
+ * all, so `InvestigationService`'s `if (env && (inputTokens || outputTokens))` was always false:
+ * `investigations.model` stayed null and Autumn was never metered for autonomous investigations,
+ * which the pre-`@maple/llm` workflow path did meter.
+ */
+export interface TurnUsage {
+	input: number
+	output: number
+	cacheRead: number
+}
+
+export const makeTurnUsage = (): TurnUsage => ({ input: 0, output: 0, cacheRead: 0 })
+
+export const addUsage = (total: TurnUsage, usage: Usage | undefined): void => {
+	total.input += usage?.inputTokens ?? 0
+	total.output += usage?.outputTokens ?? 0
+	total.cacheRead += usage?.cacheReadInputTokens ?? 0
+}
+
+/**
+ * Stamp an event with the ref that routes it into a parent's task card.
+ *
+ * Every emission site goes through this, so a sub-agent's events can never leak into the top-level
+ * conversation by omission — the tag is applied once, at the boundary, rather than remembered at
+ * each of the seven places a turn emits. Typed structurally so this module does not depend on the
+ * whole turn input.
+ */
+export const tagged = <E extends ChatTurnEvent>(source: { readonly task?: ChatTaskRef }, event: E): E =>
+	source.task === undefined ? event : { ...event, task: source.task }
+
+export const turnEnd = (
+	input: { readonly messageId: string; readonly task?: ChatTaskRef },
+	reason: Extract<ChatEvent, { type: "turn-end" }>["reason"],
+	error?: string,
+): ChatTurnEvent =>
+	tagged(input, {
+		type: "turn-end",
+		messageId: input.messageId,
+		reason,
+		...(error === undefined ? {} : { error }),
+	})
+
+/** A turn with no session attached (tests, one-shot callers) is always current. */
+export const isCurrent = (input: ChatTurnInput): boolean => input.isCurrent === undefined || input.isCurrent()
